Refactor: Moved page sections into individual components
